@@ -11,13 +11,14 @@
     "build:middleware": "cd src/lib/middleware/ && webpack",
     "build:venom": "tsc",
     "build:wapi": "cd src/lib/wapi/ && webpack",
-    "build": "npm run build:wapi && npm run build:middleware && npm run build:jsQR && npm run build:counter && npm run build:venom",
+    "build": "npm run fix:imports && npm run build:wapi && npm run build:middleware && npm run build:jsQR && npm run build:counter && npm run build:venom",
     "changelog:last": "conventional-changelog -p angular -r 2",
     "changelog:preview": "conventional-changelog -p angular -u",
     "changelog:update": "conventional-changelog -p angular -i CHANGELOG.md -s",
     "clean": "shx rm -rf session dist",
     "commit": "cz",
     "fix:doc-theme": "shx cp ./node_modules/typedoc-default-themes/bin/default/partials/type.hbs ./node_modules/typedoc-plugin-pages/dist/theme/v2/partials/type.hbs",
+    "fix:imports": "eslint --fix src/**/*.{js,jsx,ts,tsx} && npm run lint:fix",
     "generate-api-docs": "npm run fix:doc-theme && typedoc",
     "lint:js": "npx eslint -c .eslintrc.js --ext .js src",
     "lint:ts": "npx eslint -c .eslintrc.js --ext .ts src",
@@ -79,13 +80,16 @@
     "@types/mime-types": "^2.1.0",
     "@types/node": "^18.11.10",
     "@types/puppeteer": "^5.4.6",
+    "@types/qrcode-terminal": "^0.12.0",
     "@types/sharp": "^0.30.5",
     "@types/ws": "^8.5.3",
     "concurrently": "^7.6.0",
     "conventional-changelog-cli": "^2.2.2",
     "copy-webpack-plugin": "^10.2.4",
-    "eslint": "^8.32.0",
+    "eslint": "^8.37.0",
     "eslint-config-prettier": "^6.14.0",
+    "eslint-import-resolver-typescript": "^3.5.5",
+    "eslint-plugin-import": "^2.27.5",
     "eslint-plugin-prettier": "^4.2.1",
     "gulp": "^4.0.2",
     "husky": "^8.0.1",
