@@ -52,119 +52,119 @@ MMMMMMMMMMMMMMMMMMMMMMMMMMMMMMMMMMMMMMMMMMMMMMMMMMMMMMNMMNMNMMMNMMNNMMMMMMMMMMMM
 MMMMMMMMMMMMMMMMMMMMMMMMMMMMMMMMMMMMMMMMMMMMMMMMMMMMMMNNNNMMNNNMMMMMMMMMMMMMMMMMMMMMMMMMMMMMMMMMMMMM
 MMMMMMMMMMMMMMMMMMMMMMMMMMMMMMMMMMMMMMMMMMMMMMMMMMMMMMMMMMMMMMMMMMMMMMMMMMMMMMMMMMMMMMMMMMMMMMMMMMMM
 */
-export { areAllMessagesLoaded } from './are-all-messages-loaded';
-export { clearChatMessages } from './clear-chat';
-export { createGroup } from './create-group';
-export { deleteConversation } from './delete-conversation';
-export { deleteMessagesAll } from './delete-messages-all';
-export { deleteMessagesMe } from './delete-messages-me';
-export { downloadFile } from './download-file-with-credentials';
-export { encryptAndUploadFile } from './encrypt-and-upload-file';
-export { getAllChats } from './get-all-chats';
-export { getAllChatIds } from './get-all-chats-ids';
-export { getAllChatsWithMessages } from './get-all-chats-with-messages';
-export { getAllContacts } from './get-all-contacts';
-export { getAllGroupMetadata } from './get-all-group-metadata';
-export { getAllGroups } from './get-all-groups';
-export { getAllMessagesInChat } from './get-all-messages-in-chat';
-export { getAllNewMessages } from './get-all-new-messages';
-export { getAllUnreadMessages } from './get-all-unread-messages';
-export { getBatteryLevel } from './get-battery-level';
-export { getChat } from './get-chat';
-export { getChatById } from './get-chat-by-id';
-export { getChatByName } from './get-chat-by-name';
-export { getAllChatsWithNewMessages } from './get-chats-with-new-messages';
-export { getCommonGroups } from './get-common-groups';
-export { getContact } from './get-contact';
-export { getGroupAdmins } from './get-group-admins';
-export { getGroupInviteLink } from './get-group-invite-link';
-export { getGroupInfoFromInviteLink } from './get-group-info-from-invite-link';
-export { getGroupMetadata } from './get-group-metadata';
-export { getGroupParticipantIDs } from './get-group-participant-ids';
-export { _getGroupParticipants } from './get-group-participants';
-export { getHost } from './get-host';
-export { getMe } from './get-me';
-export { getMyContacts } from './get-my-contacts';
-export { getNewId } from './get-new-id';
-export { getNumberProfile } from './get-number-profile';
-export { getProfilePicFromServer } from './get-profile-pic-from-server';
-export { getStatus } from './get-status';
-export { getUnreadMessages } from './get-unread-messages';
-export { getUnreadMessagesInChat } from './get-unread-messages-in-chat';
-export { hasUndreadMessages } from './has-unread-messages';
-export { isConnected } from './is-connected';
-export { isLoggedIn } from './is-logged-in';
-export { leaveGroup } from './leave-group';
+export { areAllMessagesLoaded } from './are-all-messages-loaded.js';
+export { clearChatMessages } from './clear-chat.js';
+export { createGroup } from './create-group.js';
+export { deleteConversation } from './delete-conversation.js';
+export { deleteMessagesAll } from './delete-messages-all.js';
+export { deleteMessagesMe } from './delete-messages-me.js';
+export { downloadFile } from './download-file-with-credentials.js';
+export { encryptAndUploadFile } from './encrypt-and-upload-file.js';
+export { getAllChats } from './get-all-chats.js';
+export { getAllChatIds } from './get-all-chats-ids.js';
+export { getAllChatsWithMessages } from './get-all-chats-with-messages.js';
+export { getAllContacts } from './get-all-contacts.js';
+export { getAllGroupMetadata } from './get-all-group-metadata.js';
+export { getAllGroups } from './get-all-groups.js';
+export { getAllMessagesInChat } from './get-all-messages-in-chat.js';
+export { getAllNewMessages } from './get-all-new-messages.js';
+export { getAllUnreadMessages } from './get-all-unread-messages.js';
+export { getBatteryLevel } from './get-battery-level.js';
+export { getChat } from './get-chat.js';
+export { getChatById } from './get-chat-by-id.js';
+export { getChatByName } from './get-chat-by-name.js';
+export { getAllChatsWithNewMessages } from './get-chats-with-new-messages.js';
+export { getCommonGroups } from './get-common-groups.js';
+export { getContact } from './get-contact.js';
+export { getGroupAdmins } from './get-group-admins.js';
+export { getGroupInviteLink } from './get-group-invite-link.js';
+export { getGroupInfoFromInviteLink } from './get-group-info-from-invite-link.js';
+export { getGroupMetadata } from './get-group-metadata.js';
+export { getGroupParticipantIDs } from './get-group-participant-ids.js';
+export { _getGroupParticipants } from './get-group-participants.js';
+export { getHost } from './get-host.js';
+export { getMe } from './get-me.js';
+export { getMyContacts } from './get-my-contacts.js';
+export { getNewId } from './get-new-id.js';
+export { getNumberProfile } from './get-number-profile.js';
+export { getProfilePicFromServer } from './get-profile-pic-from-server.js';
+export { getStatus } from './get-status.js';
+export { getUnreadMessages } from './get-unread-messages.js';
+export { getUnreadMessagesInChat } from './get-unread-messages-in-chat.js';
+export { hasUndreadMessages } from './has-unread-messages.js';
+export { isConnected } from './is-connected.js';
+export { isLoggedIn } from './is-logged-in.js';
+export { leaveGroup } from './leave-group.js';
 export {
   asyncLoadAllEarlierMessages,
   loadAllEarlierMessages
-} from './load-all-earlier-chat-messages';
-export { loadAndGetAllMessagesInChat } from './load-and-get-all-messages-in-chat';
-export { loadChatEarlierMessages } from './load-earlier-chat-messages';
-export { loadEarlierMessagesTillDate } from './load-earlier-messages-til-date';
-export { processFiles } from './process-files';
-export { processMessageObj } from './process-message-object';
-export { revokeGroupInviteLink } from './revoke-invite-link';
-export { sendChatstate } from './send-chat-state';
-export { sendFile } from './send-file';
-export { sendImage } from './send-image';
-export { sendPtt } from './send-ptt';
-export { sendImageAsSticker } from './send-image-as-stricker';
-export { sendImageWithProduct } from './send-image-with-product';
-export { sendLocation } from './send-location';
-export { sendMessage } from './send-message';
-export { sendMessageOptions } from './sendMessageOptions';
-export { sendMessageWithTags } from './send-message-with-tags';
-export { sendMessageWithThumb } from './send-message-with-thumb';
-export { sendMessage2 } from './send-message2';
-export { sendSeen } from './send-seen';
-export { sendSticker } from './send-sticker';
-export { sendVideoAsGif } from './send-video-as-gif';
-export { setMyName } from './set-my-name';
-export { setMyStatus } from './set-my-status';
-export { forwardMessages } from './forward-messages';
-export { sendContactVcard } from './send-contact-vcard';
-export { getNewMessageId } from './get-new-message-id';
-export { reply } from './reply';
-export { startTyping, stopTyping } from './simulate-typing';
-export { getMessageById } from './get-message-by-id';
-export { blockContact } from './block-contact';
-export { unblockContact } from './unblock-contact';
-export { removeParticipant } from './remove-participant';
-export { addParticipant } from './add-participant';
-export { promoteParticipant } from './promote-participant';
-export { demoteParticipant } from './demote-participant';
-export { openChat, openChatAt } from './open-chat';
-export { joinGroup } from './join-group';
+} from './load-all-earlier-chat-messages.js';
+export { loadAndGetAllMessagesInChat } from './load-and-get-all-messages-in-chat.js';
+export { loadChatEarlierMessages } from './load-earlier-chat-messages.js';
+export { loadEarlierMessagesTillDate } from './load-earlier-messages-til-date.js';
+export { processFiles } from './process-files.js';
+export { processMessageObj } from './process-message-object.js';
+export { revokeGroupInviteLink } from './revoke-invite-link.js';
+export { sendChatstate } from './send-chat-state.js';
+export { sendFile } from './send-file.js';
+export { sendImage } from './send-image.js';
+export { sendPtt } from './send-ptt.js';
+export { sendImageAsSticker } from './send-image-as-stricker.js';
+export { sendImageWithProduct } from './send-image-with-product.js';
+export { sendLocation } from './send-location.js';
+export { sendMessage } from './send-message.js';
+export { sendMessageOptions } from './sendMessageOptions.js';
+export { sendMessageWithTags } from './send-message-with-tags.js';
+export { sendMessageWithThumb } from './send-message-with-thumb.js';
+export { sendMessage2 } from './send-message2.js';
+export { sendSeen } from './send-seen.js';
+export { sendSticker } from './send-sticker.js';
+export { sendVideoAsGif } from './send-video-as-gif.js';
+export { setMyName } from './set-my-name.js';
+export { setMyStatus } from './set-my-status.js';
+export { forwardMessages } from './forward-messages.js';
+export { sendContactVcard } from './send-contact-vcard.js';
+export { getNewMessageId } from './get-new-message-id.js';
+export { reply } from './reply.js';
+export { startTyping, stopTyping } from './simulate-typing.js';
+export { getMessageById } from './get-message-by-id.js';
+export { blockContact } from './block-contact.js';
+export { unblockContact } from './unblock-contact.js';
+export { removeParticipant } from './remove-participant.js';
+export { addParticipant } from './add-participant.js';
+export { promoteParticipant } from './promote-participant.js';
+export { demoteParticipant } from './demote-participant.js';
+export { openChat, openChatAt } from './open-chat.js';
+export { joinGroup } from './join-group.js';
 export { markUnseenMessage } from './mark-unseen-message.js';
-export { getBlockList } from './block-list';
-export { setTheme, getTheme } from './theme';
-export { restartService } from './restart-service';
-export { killServiceWorker } from './kill-service-worker';
-export { sendLinkPreview } from './send-link-preview';
-export { sendQueryMsgInfo } from './send-query-message-info';
+export { getBlockList } from './block-list.js';
+export { setTheme, getTheme } from './theme.js';
+export { restartService } from './restart-service.js';
+export { killServiceWorker } from './kill-service-worker.js';
+export { sendLinkPreview } from './send-link-preview.js';
+export { sendQueryMsgInfo } from './send-query-message-info.js';
 export {
   sendExist,
   scope,
   newId,
   getchatId,
   sendCheckType
-} from './check-send-exist';
-export { sendContactVcardList } from './send-contact-vcard-list';
-export { setProfilePic } from './set-profile-pic';
-export { pinChat } from './fix-chat';
-export { getSessionTokenBrowser } from './get-session-token';
-export { sendMute } from './send-mute';
-export { getListMute, interfaceMute } from './get-list-mute';
-export { downloadMedia } from './download-media';
-export { checkIdMessage } from './check-id-messagem';
-export { returnReply } from './return-reply';
-export { logout } from './logout';
-export { setGroupDescription } from './set-group-description';
-export { setGroupTitle } from './set-group-title';
-export { setGroupSettings } from './set-group-settings';
-export { sendButtons } from './send-buttons';
-export { sendListMenu } from './send-list-menu';
-export { checkChat } from './check-chat';
-export { checkNumberStatus } from './check-number-status';
-export { isBeta } from './check-beta';
+} from './check-send-exist.js';
+export { sendContactVcardList } from './send-contact-vcard-list.js';
+export { setProfilePic } from './set-profile-pic.js';
+export { pinChat } from './fix-chat.js';
+export { getSessionTokenBrowser } from './get-session-token.js';
+export { sendMute } from './send-mute.js';
+export { getListMute, interfaceMute } from './get-list-mute.js';
+export { downloadMedia } from './download-media.js';
+export { checkIdMessage } from './check-id-messagem.js';
+export { returnReply } from './return-reply.js';
+export { logout } from './logout.js';
+export { setGroupDescription } from './set-group-description.js';
+export { setGroupTitle } from './set-group-title.js';
+export { setGroupSettings } from './set-group-settings.js';
+export { sendButtons } from './send-buttons.js';
+export { sendListMenu } from './send-list-menu.js';
+export { checkChat } from './check-chat.js';
+export { checkNumberStatus } from './check-number-status.js';
+export { isBeta } from './check-beta.js';
